pyproject.toml
[tool.poetry]
name = "trading-platform"
version = "0.1.0"
description = ""
authors = ["Owwwen86 <smirnovkirill@mail.ru>"]
readme = "README.md"
packages = [{include = "trading_platform"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
generics = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
